chore: auto-derive pa11y-ci URL list from filesystem walk
The previous .pa11yci.json hand-listed 36 URLs and had drifted out of
sync with the actual page set — none of the 30+ generated examples,
no lecture pages, and only a handful of exercises were covered
(issue #394).

Replace it with .pa11yci.js that reuses the same directory walk as
scripts/build-sitemap.js (now extracted to scripts/collect-html.js)
to discover every page automatically. Full-scan coverage grows from
36 to 171 URLs and stays current as pages are added.

Concurrency bumped to 4 to keep CI runtime bounded. The existing
partial-scan logic in checks.yml still applies on PRs, so most runs
remain fast; only direct pushes to master and shared-asset PRs
trigger the full scan.
