+++ .github/workflows/checks.yml
@@ -64,7 +64,7 @@ jobs:
                   echo "Running a11y on ${#URLS[@]} changed file(s):"
                   printf '  %s\n' "${URLS[@]}"
                   npx start-server-and-test 'npm run serve' http://localhost:8000 \
-                    "npx pa11y-ci --config .pa11yci.json ${URLS[*]}"
+                    "npx pa11y-ci --config .pa11yci.js ${URLS[*]}"
 
     typos:
         name: Spell check (typos)
