--- .pa11yci.json
@@ -1,47 +1,0 @@
-{
-	"defaults": {
-		"standard": "WCAG2AA",
-		"timeout": 30000,
-		"wait": 500,
-		"chromeLaunchConfig": {
-			"args": ["--no-sandbox"]
-		}
-	},
-	"urls": [
-		"http://localhost:8000/index.html",
-		"http://localhost:8000/course/index.html",
-		"http://localhost:8000/course/COBOLIntro.html",
-		"http://localhost:8000/course/COBOLcommands.html",
-		"http://localhost:8000/course/Copy.html",
-		"http://localhost:8000/course/DataDeclaration.html",
-		"http://localhost:8000/course/EditedPics.html",
-		"http://localhost:8000/course/IndexedFiles.html",
-		"http://localhost:8000/course/Inspect.html",
-		"http://localhost:8000/course/Intro2DirectAccess.html",
-		"http://localhost:8000/course/Iteration.html",
-		"http://localhost:8000/course/RefMod.html",
-		"http://localhost:8000/course/RelativeFiles.html",
-		"http://localhost:8000/course/ReportWriter.html",
-		"http://localhost:8000/course/ReportWriterSS.html",
-		"http://localhost:8000/course/Search.html",
-		"http://localhost:8000/course/Selection.html",
-		"http://localhost:8000/course/SequentialFiles1.html",
-		"http://localhost:8000/course/SequentialFiles2.html",
-		"http://localhost:8000/course/SequentialFiles3.html",
-		"http://localhost:8000/course/SortMerge.html",
-		"http://localhost:8000/course/String.html",
-		"http://localhost:8000/course/Subprograms.html",
-		"http://localhost:8000/course/Tables1.html",
-		"http://localhost:8000/course/Tables2.html",
-		"http://localhost:8000/course/Unstring.html",
-		"http://localhost:8000/course/Usage.html",
-		"http://localhost:8000/exercises/index.html",
-		"http://localhost:8000/exercises/CountDown.html",
-		"http://localhost:8000/exercises/COBOLExams/COBOLExams.html",
-		"http://localhost:8000/exercises/Exm-AcmeStockReorder/Exm-AcmeStockReorder.html",
-		"http://localhost:8000/exercises/Exm-AcmeStockReorder/Prg-AcmeStockReorder.html",
-		"http://localhost:8000/exercises/Prj-NewtronicsFileMaint/Prj-NewtronicsFileMaint.html",
-		"http://localhost:8000/exercises/Proj-CSISEvalform/CSISEvalForm.html",
-		"http://localhost:8000/examples/default.html"
-	]
-}
