+++ package.json
@@ -11,7 +11,7 @@
 		"check:img-dims": "node scripts/check-img-dimensions.js",
 		"links:run": "linkinator http://localhost:8000 --config linkinator.config.json",
 		"links": "start-server-and-test serve http://localhost:8000 links:run",
-		"a11y:run": "pa11y-ci --config .pa11yci.json",
+		"a11y:run": "pa11y-ci --config .pa11yci.js",
 		"a11y": "start-server-and-test serve http://localhost:8000 a11y:run",
 		"check:assets": "node scripts/check-assets.js",
 		"format:check": "prettier --check .",
